Backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["defalt91 <armansoltanian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
SQLAlchemy = { extras = ["mypy"], version = "^1.4.26" }
slowapi = "^0.1.5"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
uvicorn = "^0.15.0"
email-validator = "^1.1.3"
emails = "^0.6"
python-multipart = "^0.0.5"
tenacity = "^8.0.1"
starlette-csrf = "^1.3.0"
asyncpg = "^0.24.0"
aiopg = "^1.3.2"
alembic = "^1.7.4"
Jinja2 = "^3.0.2"
sqlalchemy2-stubs = "^0.0.2-alpha.19"
Authlib = "^0.15.5"
timing-asgi = "^0.2.1"
starlette-prometheus = "^0.8.0"
httpx = "^0.20.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.black]
